+++ README.md
@@ -120,3 +120,4 @@ O[RookStrategy] --> J
 - [x] rank가 1, file이 2이면 움직일 수 있다.
 - [x] rank가 2, file이 1이 아닐 때 움직일 수 없다.
 - [x] rank가 2, file이 1일 때 움직일 수 있다.
+
