+++ src/main/java/chess/ui/OutputView.java
@@ -1,6 +1,6 @@
 package chess.ui;
 
-import chess.domain.dto.response.PiecesResponse;
+import chess.dto.response.PiecesResponse;
 
 import static chess.domain.Pieces.FIRST_FILE_OF_BLACK;
 import static chess.domain.Pieces.FIRST_FILE_OF_WHITE;
